SocialService-Refactoring: Moved implemented controllers to a new package.
